src/main/java/com/nojong/book/springboot/web/HelloController.java
package com.nojong.book.springboot.web;

import com.nojong.book.springboot.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

/*
*   @RestController
*   - Controller를 JSON을 반환하는 컨트롤러로 만들어 준다.
*   - @Controller + @ResponseBody = @RestController
* */
@RestController
public class HelloController {

    @GetMapping("/hello")
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name,
                                     @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }
}
